+++ extend/waibuMpa/route/user/activation.js
@@ -10,7 +10,9 @@ const userActivation = {
         const query = { status: 'UNVERIFIED', token: req.body.key }
         const result = await model.findRecord({ query, limit: 1 })
         if (result.length === 0) throw this.error('validationError', { details: [{ field: 'key', error: 'invalidActivationKey' }] })
-        await model.updateRecord(result[0].id, { status: 'ACTIVE' }, { req, noValidation: true, noFlash: true })
+        await model.transaction(async (trx) => {
+          await model.updateRecord(result[0].id, { status: 'ACTIVE' }, { req, noValidation: true, noFlash: true, trx })
+        })
         req.flash('notify', req.t('userActivated'))
         return reply.redirectTo(this.config.redirect.signin, req)
       } catch (err) {