+++ extend/waibuMpa/route/user/signup.js
@@ -24,7 +24,7 @@ const signup = {
         const validation = { ns: ['sumba', 'dobo'], fields, extFields }
         req.body.token = generateId()
         req.body.provider = 'local'
-        const { data } = await createRecord({ model: 'SumbaUser', req, reply, options: { validation, noFlash: true, forceNoHidden: true } })
+        const { data } = await createRecord({ model: 'SumbaUser', req, reply, transaction: true, options: { validation, noFlash: true, forceNoHidden: true } })
         req.flash('notify', req.t('userCreated'))
         return await reply.view('sumba.template:/user/signup/success.html', { form: req.body, data })
       } catch (err) {